Stable version of import - was able to complete a run of fundamentals for both NYSE and NASDAQ without serialization errors.
Made many tweaks along the way to deal with the errors I got out of my real-world tests of importing fundamentals.

Removed api_responses from the database and removed the relevant code; going to rethink how api request/responses are preserved.

Removed has_options from symbols. Decided to handle this differently.

+++ src/EodHistoricalData/NuGet/EodHistoricalData.Sdk/Models/Fundamentals/CommonStock/EsgScores.cs
@@ -10,11 +10,11 @@ public struct EsgScores
     public double? TotalEsg;
     public double? TotalEsgPercentile;
     public double? EnvironmentScore;
-    public int? EnvironmentScorePercentile;
+    public double? EnvironmentScorePercentile;
     public double? SocialScore;
-    public int? SocialScorePercentile;
+    public double? SocialScorePercentile;
     public double? GovernanceScore;
-    public int? GovernanceScorePercentile;
-    public int? ControversyLevel;
+    public double? GovernanceScorePercentile;
+    public double? ControversyLevel;
     public IDictionary<string, ActivityInvolvement>? ActivitiesInvolvement;
 }
